Izin hesabi: haftalik izin 1 gun (Pazar), yarim gun resmi tatil, bucuklu gun
- businessDays: yalnizca Pazar haftalik izin sayilir; Cumartesi calisma gunu
- Tam gun resmi tatil izinden sayilmaz; yarim gun resmi tatil (arife) 0,5 gun
- fullHolidays / halfHolidays listeleri (2026 dini bayramlar dahil - dogrulanacak)
- leave_requests.days integer -> numeric(5,1) (migration 028)
- POST /api/leaves: gun sayisi 0,5 katlari kabul eder
- Kalan/kullanilan izin ve talep sureleri bucuklu gosterilir (leaveNumber)

+++ api/init.sql
@@ -276,7 +276,7 @@ create table if not exists leave_requests(
   leave_type text not null,
   start_date date not null,
   end_date date not null,
-  days integer not null check(days > 0),
+  days numeric(5,1) not null check(days > 0),
   status text not null default 'Bekliyor',
   approval_route jsonb not null default '[]'::jsonb,
   approval_step integer not null default 0,
